+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/runtime/builtins/temporal/JSTemporalPlainDateTimeObject.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -47,20 +47,20 @@ import com.oracle.truffle.js.runtime.objects.JSNonProxyObject;
 public class JSTemporalPlainDateTimeObject extends JSNonProxyObject implements TemporalMonth, TemporalYear, TemporalDay, TemporalCalendar {
 
     // from time
-    private final long hours;
-    private final long minutes;
-    private final long seconds;
-    private final long milliseconds;
-    private final long microseconds;
-    private final long nanoseconds;
+    private final int hours;
+    private final int minutes;
+    private final int seconds;
+    private final int milliseconds;
+    private final int microseconds;
+    private final int nanoseconds;
     // from date
     private final int year;
     private final int month;
     private final int day;
     private final DynamicObject calendar;
 
-    protected JSTemporalPlainDateTimeObject(Shape shape, long year, long month, long day, long hours, long minutes, long seconds, long milliseconds,
-                    long microseconds, long nanoseconds, DynamicObject calendar) {
+    protected JSTemporalPlainDateTimeObject(Shape shape, int year, int month, int day, int hours, int minutes, int seconds, int milliseconds,
+                    int microseconds, int nanoseconds, DynamicObject calendar) {
         super(shape);
         this.hours = hours;
         this.minutes = minutes;
@@ -69,48 +69,48 @@ public class JSTemporalPlainDateTimeObject extends JSNonProxyObject implements T
         this.microseconds = microseconds;
         this.nanoseconds = nanoseconds;
 
-        this.year = (int) year;
-        this.month = (int) month;
-        this.day = (int) day;
+        this.year = year;
+        this.month = month;
+        this.day = day;
         this.calendar = calendar;
     }
 
-    public long getHour() {
+    public int getHour() {
         return hours;
     }
 
-    public long getMinute() {
+    public int getMinute() {
         return minutes;
     }
 
-    public long getSecond() {
+    public int getSecond() {
         return seconds;
     }
 
-    public long getMillisecond() {
+    public int getMillisecond() {
         return milliseconds;
     }
 
-    public long getMicrosecond() {
+    public int getMicrosecond() {
         return microseconds;
     }
 
-    public long getNanosecond() {
+    public int getNanosecond() {
         return nanoseconds;
     }
 
     @Override
-    public long getYear() {
+    public int getYear() {
         return year;
     }
 
     @Override
-    public long getMonth() {
+    public int getMonth() {
         return month;
     }
 
     @Override
-    public long getDay() {
+    public int getDay() {
         return day;
     }
