+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/runtime/builtins/temporal/TemporalTime.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * The Universal Permissive License (UPL), Version 1.0
@@ -41,15 +41,15 @@
 package com.oracle.truffle.js.runtime.builtins.temporal;
 
 public interface TemporalTime extends TemporalCalendar {
-    long getHour();
+    int getHour();
 
-    long getMinute();
+    int getMinute();
 
-    long getSecond();
+    int getSecond();
 
-    long getMillisecond();
+    int getMillisecond();
 
-    long getMicrosecond();
+    int getMicrosecond();
 
-    long getNanosecond();
+    int getNanosecond();
 }
